# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Business Analytics Dashboard API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /sales:
    get:
      operationId: getSales
      tags: [sales]
      summary: List sales records
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
        - name: search
          in: query
          schema:
            type: string
        - name: category
          in: query
          schema:
            type: string
        - name: channel
          in: query
          schema:
            type: string
        - name: dateFrom
          in: query
          schema:
            type: string
        - name: dateTo
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Paginated sales records
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SalesList"
    post:
      operationId: createSale
      tags: [sales]
      summary: Create a new sale record
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SaleInput"
      responses:
        "201":
          description: Created sale
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Sale"
...
  /sales/trend:
    get:
      operationId: getSalesTrend
      tags: [sales]
      summary: Sales trend chart data
      parameters:
        - name: period
          in: query
          schema:
            type: string
            default: daily
      responses:
        "200":
          description: Sales trend
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ChartDataPoint"
...
  /sales/{id}:
    put:
      operationId: updateSale
      tags: [sales]
      summary: Update a sale record
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SaleUpdate"
      responses:
        "200":
          description: Updated sale
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Sale"
    delete:
      operationId: deleteSale
      tags: [sales]
      summary: Delete a sale record
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeleteResult"

  # ─── INVENTORY ────────────────────────────────────────────────
...
components:
  schemas:
...
    DeleteResult:
      type: object
      required: [success]
      properties:
        success:
          type: boolean
    ChartDataPoint:
      type: object
      required: [label, value]
      properties:
        label:
          type: string
        value:
          type: number
        secondary:
          type: ["number", "null"]
...
    Sale:
      type: object
      required: [id, invoiceId, date, customer, product, category, qty, unitPrice, revenue, channel]
      properties:
        id:
          type: integer
        invoiceId:
          type: string
        date:
          type: string
        customer:
          type: string
        product:
          type: string
        category:
          type: string
        qty:
          type: integer
        unitPrice:
          type: number
        revenue:
          type: number
        channel:
          type: string
    SaleInput:
      type: object
      required: [customer, product, category, qty, unitPrice, channel]
      properties:
        customer:
          type: string
        product:
          type: string
        category:
          type: string
        qty:
          type: integer
        unitPrice:
          type: number
        channel:
          type: string
        date:
          type: ["string", "null"]
    SaleUpdate:
      type: object
      properties:
        customer:
          type: string
        product:
          type: string
        category:
          type: string
        qty:
          type: integer
        unitPrice:
          type: number
        channel:
          type: string
    SalesList:
      type: object
      required: [data, total, page, limit]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Sale"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
...