# lib/api-zod/src/generated/types/salesList.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Business Analytics Dashboard API
 * OpenAPI spec version: 0.1.0
 */
import type { Sale } from "./sale";

export interface SalesList {
  data: Sale[];
  total: number;
  page: number;
  limit: number;
}